requirements.txt
numpy
matplotlib
soundfile
ipykernel
pandas
python-sofa
torch
torchaudio
scipy
tqdm
scaper
speechbrain
lightning
wandb
asteroid
pyyaml
GitPython
resemblyzer
espnet
cdpam
auraloss
seaborn
ptflops
requests
openai-whisper
dotenv
